server/requirements.txt
# botocore==1.35.81
# aiobotocore==2.16.0
# boto3==1.35.81
boto3
awscli
aiobotocore[boto3]
aioboto3
aiohttp
aiosignal
annotated-types
anyio
appnope 
asttokens 
attrs
bidict
certifi
charset-normalizer
click 
comm 
contourpy
cycler
dataclasses-json
debugpy 
decorator 
distro
executing 
faiss-cpu
fastjsonschema
Flask
Flask-Cors
Flask-SocketIO
fonttools
frozenlist
h11
httpcore
httpx
idna
ipykernel 
ipython 
itsdangerous
jedi 
Jinja2 
joblib 
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client 
jupyter_core
kiwisolver
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
multidict
mypy-extensions
nbformat
nest-asyncio
nltk
numpy
openai
orjson
packaging
pandas
parso
pexpect
pillow
platformdirs
plotly
prompt-toolkit
psutil
ptyprocess
pure-eval
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
python-engineio
python-socketio
pytz
PyYAML
pyzmq
referencing
regex
requests
rpds-py
scikit-learn
scipy
seaborn
setuptools
simple-websocket
six
sniffio
SQLAlchemy
stack-data
tenacity
threadpoolctl
tiktoken
tornado
tqdm
traitlets
typing-inspect
typing_extensions
tzdata
urllib3
wcwidth
Werkzeug
wheel
wsproto
yarl
firebase_admin
mermaid-py
awscli